chore: externalize code for creating simple geometry
